Modification for STM32 CubeHAL defect.
The CubeHAL API HAL_I2S_Receive_DMA() and HAL_I2S_Transmit_DMA() have
probelem. While the HAL manual says the length of the data have to be
given as the length by int16, the API interpret it as length by int32,
when the I2S frame type is 32bit or 24bit.

So, the compensation part is disabled at this moment.