24 bit I2S error is fixed.
Issue #97 : 24bit I2S data is not correctly scaled.
https://github.com/suikan4github/murasaki/issues/97

Added 8bit shift to the DMA data in case of the 24bit I2S. This was
required because the 24bit data is handled as integer in the STM32
architecture.

Noet that the I2S data have to be handled as signed integer, but the DMA
fills upper 8bit as zero always. In this fix, this point is cared.