Fix P2/P3 overflow.
In the certain case, the P3 or P2 excced
the 2^20 which is the limit of Si5351.

To prevent it, the numerator and denominator
of the PLL divider fractional part are
scaled down until P2 and P3 avoid the
overflow.

This may introduce the frequency error.
Theoritically, it is less than 0.05ppm.
( The integer part is at least 24 when
the XTAL is 25MHz. Thus, the error is
less thatn 1 / 24 * 2^20