Update button switch definition in template.
The nucle button swithc is named B1 usually. But sometimes it is
depend on board.

As as result, name is difficut to set. Some Nucleo comes with EXTI enabled,
Some Nucleo comes with EXIT disabled. After confused work, we decided to set
it default user button swtchi name as B1 what ever it is EXTI or not.

The name is refered in the murasaki_sample repository onry. Other application
can change the name freely.